migrate a bunch of code to cairo drawing
A lot, but not all of gdk_draw_drawable() calls have been replaced by a
gkrellm_draw_drawable() wrapper that so far simplified the migration and
seems to work.
Migrating panel.c is not as straight forward as the other files, so this
commit is used as a save point for additional work.